[REFACTOR/#485] enum package -> enums 네이밍 변경
