+++ app/src/main/java/com/el/yello/presentation/main/myyello/MyYelloAdapter.kt
@@ -9,7 +9,7 @@ import androidx.recyclerview.widget.RecyclerView
 import com.el.yello.R
 import com.el.yello.databinding.ItemMyYelloBinding
 import com.example.domain.entity.Yello
-import com.example.domain.enum.Gender
+import com.example.domain.enums.Gender
 import com.example.ui.extension.setOnSingleClickListener
 import com.example.ui.util.Utils.setChosungText
