+++ domain/src/main/java/com/example/domain/entity/PayUserSubsInfoModel.kt
@@ -1,6 +1,6 @@
 package com.example.domain.entity
 
-import com.example.domain.enum.SubscribeType
+import com.example.domain.enums.SubscribeType
 
 class PayUserSubsInfoModel(
     val id: Long,
