+++ domain/src/main/java/com/example/domain/enums/Gender.kt
@@ -1,4 +1,4 @@
-package com.example.domain.enum
+package com.example.domain.enums
 
 enum class Gender {
     M, W;
