+++ domain/src/main/java/com/example/domain/enums/PointEnum.kt
@@ -1,4 +1,4 @@
-package com.example.domain.enum
+package com.example.domain.enums
 
 enum class PointEnum {
     INITIAL, KEYWORD, SUBSCRIBE
